agent_b_caller/requirements.txt
a2a-sdk
python-dotenv>=1.0.0
uvicorn[standard]>=0.30.0
starlette>=0.37.0
httpx>=0.27.0